requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cycler==0.11.0
fire==0.4.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
hydra-core==1.2.0
idna==3.3
importlib-metadata==4.11.4
iniconfig==1.1.1
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.5.2
multidict==6.0.2
numpy==1.21.6
oauthlib==3.2.0
omegaconf==2.2.1
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
pytorch-lightning==1.6.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
tomli==2.0.1
torch==1.8.1
torchaudio==0.8.1
torchmetrics==0.8.2
torchtyping==0.1.4
torchvision==0.9.1
tqdm==4.64.0
typeguard==2.13.3
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
